requirements.txt

django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
whitenoise==6.7.0
